requirements.txt
flask
requests
httpx
dnslib
flask-socketio
# for WebSocket support
eventlet
# async worker for Socket.IO
flasgger
flask-cors
python-dotenv
